Input: new Collector concept and Interface.
A collector is an interface that aim to let Input plugins to define
callbacks that needs to be triggered at certain interval of times for
data collection (e.g: CPU Input Plugin).

This patch implements the registration interface and introduce the
Engine that will take care of fire and handle the events (WIP).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>