IO: new flb_io interface to handle network I/O operations
This patch implements a new interface on Fluent Bit called flb_io. It
introduce the concept of 'upstream' servers for output plugins that
requires to write data in a safer maner.

Now every plugin that flag it self with FLB_OUTPUT_TCP or FLB_OUTPUT_TLS,
the flb_output handler will create an 'upstream' context for that plugin,
then every time a plugin requires to perform a write() or read() can use
the upstream context with the flb_io_write() or flb_io_read() operations.

Basically I am moving the connection handling data out from the output
plugins and putting a robust interface for that purpose.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>