IO: moving coroutines handler into the new flb_io
The co-routines implementation is still under development, as of this patch
I am testing a new design which tries to makes a 'clean' implementation of
coroutines creation and restore point from the main event loop.

The idea of this patch is that once the output plugin invokes flb_io_write(), this
call creates a co-routine to perform the write(2) operations.

Not a definitive approach but moving forward.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>