IO: coroutine handling connections OK
This patch makes the io_write operation try to reconnect to the
backend server, it yields and return back the control to the
main event loop.

The event loop once it catch a FLB_ENGINE_EV_THREAD event, it
resumes the thread (or coroutine) in question.

== Problem to fix ==

there is an error where the caller (output plugin) is returning
before the connection is alive, we should defer that until the
write operation is complete, otherwise we may generate some troubles
on the output plugin:

  $ bin/fluent-bit -i cpu -o fluentd:12224 -V -f 2
  Fluent-Bit v0.2.0
  Copyright (C) Treasure Data

  [2015/09/16 16:10:34] [ info] Configuration
   flush time     : 2 seconds
   input plugins  : cpu
   collectors     :
  [2015/09/16 16:10:34] [ info] starting engine
  [2015/09/16 16:10:34] [debug] [upstream] connecting to 224:12224
  [2015/09/16 16:10:34] [error] Cannot connect to 224 port 12224
  [2015/09/16 16:10:34] [ warn] [upstream] could not connect to 224:12224
  [2015/09/16 16:10:34] [debug] [in_cpu] CPU 1.50% (buffer=0)
  [2015/09/16 16:10:35] [debug] [io] trying...
  [2015/09/16 16:10:35] [debug] [io] try to reconnect
  [2015/09/16 16:10:35] [debug] [io] thread has finished  <===== FIXME! ======
  [2015/09/16 16:10:35] [debug] [fluentd] ended write(2)
  [2015/09/16 16:10:35] [debug] [in_cpu] CPU 4.50% (buffer=0)
  [2015/09/16 16:10:35] [debug] [engine] resuming thread
  [2015/09/16 16:10:35] [debug] [io] connection OK

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>