Engine: improve shutdown process, give the engine more time
Before this patch, the engine when received the message FLB_ENGINE_STOP,
performed a cleanup of resources and stopped immediately.

This patch adds an improvement and make the shutdown process work in the
following way:

 - The caller triggers a stop, e.g: flb_lib_stop(...).
 - The engine now adds a 5 seconds graceful time to let the event loop
   process any pending event.
 - After the 5 seconds, the service is stopped.

This behavior is useful for output plugins that requires network IO to
finish processing data.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>