stats: new architecture to collect statistics (details below)
This patch introduce a new architecture to collect statistics from I/O plugins,
from now it uses a worker thread to collect stats updates from each I/O plugin
caller that collect events or consume some bytes:

                  +------------------------+
                  |                        |
                  |   Fluent Bit Engine    |
                  |                        |
                  +---+----------------+---+
                      |                |
          +-----------+----+    +------+--------+
          |                |    |               |
          |  Input Plugin  |    | Output Plugin |
          |                |    |               |
          +--------+-------+    +---+-----------+
                   |                |
                   | pipe(2)        | pipe(2)
                   v                |
  +----------------+-----+        +-v---------------------+
  |time, nbytes, n_events|        |time, n_bytes, n_events|
  +----------------+-----+        +-+---------------------+
                   |                |
                   |                v
               +---v----------------+--------+
               |                             |
               |  Stats collector Worker     |
               |                             |
               +-----------------------------+

For each interface that can register some statistics data, the interface
creates a pipe(2), so the I/O layer reports the data over these pipes
using the stats functions, then in the other end of the pipe there is a
thread running an event loop reading pipe data. Upon an event received,
it knows who's the owner of the data and update a local table in memory
with such information.

The good thing about pipes, is that they can grow up to 65KB and allow
us to avoid locking mechanisms to transfer the info. Every stats message
reported by the plugins is around 24 bytes, so we are pretty good on this.

Next series of patch will dispatch the statistics to a unix socket in JSON
format so it can be consumed by third party services.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>