stats: new unix domain socket server (details below)
This patch implements a new Unix domain socket server, once Fluent Bit
starts, it will create an entry at '/tmp/fluentbit.sock' and the stats
worker thread implemented in the previous patches will handle the
acceptance and contexts associated to the clients.

The objective of this server, is to be used as a delivery mechanism for
the Fluent Bit statistics, it means that for every connected client
it will write a statistics report in plain JSON format every a certain
time interval.

note: I will 'try' to make a UI to consume this data in the next series
of patches.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>