out_td: support raw msgpack data, dup memory buffer.
On some cases, the incoming msgpack data may not have the original format
expected by Fluent Bit, it can be a direct Map of fields including already
the time field.

This patch makes the Treasure Data output plugin to accept non-array data
and trust in the caller; it duplicate the memory buffer and dispatch the
content to the cloud server.

The common use case of this is in the library mode where the caller just
set a JSON message, e.g: {"time": 12345678, "key": "val"}.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>