thread: use Valgrind macros to register new stacks on co-routines
When debugging with Valgrind, it will alert about stack size changes when
a co-routine is created and a context switch is made, this usually happens
when Fluent Bit perform an 'output -> write' operation with the output plugin.

To avoid this problem, this patch makes the flb_thread interface to use the
Valgrind macros to register the new stack. This feature is only enabled if
Fluent Bit have been configured with the -DWITH_VALGRIND=On CMake option.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>