lib: new library context and create/destroy handlers (details below)
This patch introduce a new library context that aims to keep reference
of all reasources used by the internal lib mode, e.g: event loop, event
channels and others.

Now when the lib it's initialized, it creates a context, as well when it
ends, now is possible to use a new API function called:

    void flb_lib_exit(struct flb_config *config);

It destroy all resoures associated to a initalized lib, it must be used
to avoid memory leaks.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>