in_cpu: new results in percentage and improvements (details above)
This patch is a new major improvement on how the CPU usage is calculated,
it removes some useless information (for the end user) and provide specific
metrics to focus on:

 - overall CPU usage in the system, user space and kernel space
 - CPU usage per CPU core, user space and kernel space

The overall CPU usage estimate the usage in a range from a 0 to 100% of the
capacity considering all CPU cores. The CPU cores calculation estimate it
usage as it sole, not considering other cores, e.g:

 Overall => cpu[all] all=4.250000 user=2.750000 system=1.500000
   core0 => cpu[i=0] all=9.000000 user=5.000000 system=4.000000
   core1 => cpu[i=1] all=5.000000 user=4.000000 system=1.000000
   core2 => cpu[i=2] all=1.000000 user=1.000000 system=0.000000
   core3 => cpu[i=3] all=3.000000 user=2.000000 system=1.000000

Snapshots are taken every second, this patch also fix some data types to
store the info, create string buffer caches for the MsgPack report to reduce
performance penalties.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>