http_client: new HTTP client interface (details below)
Some plugins such as TD (Treasure Data) and ES (Elasticsearch) requires
to deliver the records to an HTTP end point, as of now both plugins
are using a raw TCP/TLS connection and composing their own HTTP request,
leading to do not process status responses.

This patch implement a simple and new HTTP Client interface which aims
to make HTTP requests easier; it needs to be extended for 'headers'
composing and parser support in the next patches.

> This interface can be used by input or output plugins.

note: Curl have not been used as it too heavy for what we really need,
also this new interface is friendly with our co-routines allowing to
suspend and resume at any time while some I/O is performed.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>