upstream: re-architected Upstream TCP handler (details below)
Currently an 'upstream' instance represents a context used by
output plugins to deliver some content to destinations. Per output
plugin (and if required), an upstream context existed. This works
fine for the scenario when the flush time is extended (few seconds)
but it may not be helpful when wanted to deliver data in real-time.

Since real-time is not yet a feature, the Upstream interface needed
a new design and implementation.

This patch modify the full core where now an Upstream interface
represents an end-point to deliver data that knows to to transport
either through TCP or TLS. But in addition now, the plugins inside
the 'flush' callback, requires to 'get' an 'upstream connection'.

An upstream connection now represents the TCP channel, this allows
to have multiple of them and let the parent Upstream interface to
act as a handler of a pool of connections.

note: on this scenario real-time means 'deliver data as soon as it's
received', not after a timeout.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>