in_serial: add support for a string separator (details below)
Some serial data communication have an undeterminated length, despites
for some cases the 'min_bytes' configuration option work for some use
cases, there is others where is hard to determinate where the original
message ends.

This patch implements a new configuration key 'Separator' which allows
to set a separator pattern that Fluent Bit will use to split the messages.
This parameter can be set in the INPUT configuration section, e.g:

[INPUT]
    Name      serial
    Tag       my_data
    Separator ;;

So an incoming string like: a;;b;;c;; will be processed as 3 entries:

 - a
 - b
 - c

This patch have not been tested with a 'real' FTDI interface, so it
needs to be considered experimental until further verification.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>