buffer: new [optional] interface for buffering in the file system.
This patch introduce the structure for buffering support. The current
structure design aims to perform record buffering before to starting
the flush process with the output plugins.

For this two parameters are required, the buffer_path and the buffer_workers,
where workers are a fixed number of POSIX threads that take care of I/O
operations.

This new feature is still work in process and disabled by default in the
build system.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>