buffer: implement mask_id for chunk routes (details below)
When doing buffering is critical to avoid contention. The buffering system
uses a one or more thread workers to receive instructions for store a buffer
chunk in the 'incoming' queue, but then we require to know where these
chunk should be routed.

Buffer chunks exists in thread workers context, to determinate the routes
and 'avoid' to implement mutual exclusion to access some shared data this
patch implement 'masks_id', the concept is very simple:

 - Eeach 'output instance' receives a mask number (e.g: 1, 2, 4, 8, 16...)
 - Before to dispatch a chunk request to the workers, the output route
   is determinated, so if the chunk goes to output instances 1 and 4, the
   bit mask will be '5' and this value is set in an uint64_t variable.
 - When the worker receives a chunk request, it can easily determinate
   which output instance should handle it, so it moves the chunk to the
   outgoing queue and creates the proper references in the file system

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>