engine: new signal mechanism for tasks/threads and cleanups
This patch implements signaling through the main event loop for the
tasks running under the engine, it allow each Task to indicate when
it finished.

Also a new routine exists for output plugins so when they finish,
they can return an OK or Error value, this is done through the event
loop (note that co-routines cannot return a value).

The changes introduced are affecting how routing is working at the
moment, so it can be considered work in process, some functionality
is broken as multiple destinations.

note: outpug plugins now requries
Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>