lib: msgpack: add persistent bytes parsed to unpacker struct
For buffering optimization, we always need to know the total number of bytes
parsed upon a MSGPACK_UNPACK_SUCCESS after a call to msgpack_unpacker_next().

The new field 'last_parsed' always get the value of 'parsed' before this one is
reset. For more details about this hack please refer to the following Github issue:

  https://github.com/msgpack/msgpack-c/issues/514

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>