buffer: signaling, qchunks, engine updates and more
this patch touch several parts of Fluent Bit core and merge all
required changes for buffering:

 - engine signaling
 - buffering: core & qchunks
 - input instances

Found that some extra work is required to address some issues when
enqueing the Buffer data into the Engine and tasks. Addressing that
allowing the Buffer interface to have a 'pseudo input instance'
so Tasks can be created.

More cleanups are required, work in process.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>