log: new internal logging mechanism
Before this patch the logging interface used to format strings and
printf to STDERR. But with the introduction of Thread workers for
buffering the logging calls becomes corrupted.

This patch introduce a new logging mechanim which is safe to use
from a parent process and from any worker thread (POSIX thread).

Now a central Log-worker exists to collect messages from different
places. The messages are distributed through pipes and collected
in a central place.

To-do:
 - redirect logging to a file
 - replace perror('abc') calls with new flb_errno(...) interface
 - optionally emit logging lines as event records

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>