thread: abstract interface, provide custom reference for user
Previous this patch, the coroutines implementation (thread) was
aware about specific details of the output interface, making it
hard to reuse the calls for something different than 'output'.

The following patch makes an abstraction in the Thread interface
making it agostic, so now it can be used from a different context
than output plugins.

note: this improvement is to allow coroutines in the input
interface (work in process).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>