input: add optional co-routines support for input plugins
Before this patch, the input plugins were not able to use the I/O
API provided by Fluent Bit for async networking operations as
output plugins were doing (upstreams, http, tcp).

Now an input plugin in it flags can request to operate under
a thread context (co-routine) so it will be able to perform async I/O
yielding and resuming from a custom context and stack created.

This is useful for plugins that needs to 'connect' and operate with
upstream connections.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>