proxy: new interface to load external plugins (shared libs)
The new proxy interface allows fluent bit to load a plugin
based in a shared library, as of now it only support
output plugins.

A proxy-like interface is required as some plugins might be
implemented through a different language, e.g: golang, so
some translation between callbacks and parameters is required.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>