input: fix wrong thread(s) deletion (fix #135)
The current input plugins when running in a co-routine, was doing a
wrong thread deletion, instead of delete the thread associated with
a specific ID, it was deleting 'all threads contexts', leading to
a segfault when using in_health plugin.

This patch let the input thread deletion just focus on the thread
associated to the given id.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>