sched: retry: add invalidation handler (#127)
When a task fails, likely the Engine will ask the Scheduler to
make a retry, for hence a 'Scheduler request' and a 'Task Retry'
contexts are created.

If the retry cannot be scheduled (e.g: not enough file descriptors),
the retry will be dropped. Before this patch a retry could be deleted
without care about a 'Scheduler request', so when the scheduler trigger
it requirements, it get's an invalid reference.

This patch adds a new function to the scheduler to invalidate any
request based on the opaque data type 'data', e.g:

  request->data == retry

So when a retry is going to be destroyed, it now invokes the invalidation
routine.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>