in_tail: fix wrong handling of inotify file descriptor
When monitoring files with Inotify, a file descriptor is
created, but this file descriptor is only valid for the
Inotify instance and it don't have any relationship with
common file descriptors of the operating system.

This patch remove the wrong close() over inotify file descriptor,
only inotify_rm_watch() is required.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>