filter_grep: new 'grep' filter (read below)
This patch implements the new filter_grep plugin. It implements two basic
features that can be used in conjuntion as an ACL:

 - regex
 - exclude

This plugin is not complete, as it needs to compose the new output buffer
once filtering is done. As of now it reads the configuration, filter records
and it's ready to take some action (pass or discard).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>