filter_kubernetes: use namespace and pod name as a key pair for hash key
The behavior before this patch, was that the hashing table was using the
records Tag as the hash key for the meta data cache. This generated
several entries in the cache list since multiple Tags can hold the same
meta data since they are under the same namespace and pod name.

This patch fix the old behavior and use a pair of namespace and pod name
to store the cache.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>