in_stdin: experimental support for 'parser'
This patch implements experimental support for parsers. If a parsers
file is defined and the 'parser' property set (as in in_tail), the
plugin will try to convert the message from unstructured to structured.

This functionality assumes the data is concatenated without a separator
like \n.

note: it needs a lot of testing

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>