parser: add support for fractional seconds in Regex parser
This patch adds fractional seconds resolution when parsing a time
field.

The original problem is that strptime(2) do not support fractional
seconds, other languages such as Python or Ruby implement their own
strptime()-like function to workaround the problem.

The implementation as in the other languages is a hack. In order to
perform a full resolution of the time with fractional seconds the
parser expect the following format:

  %H:%M:%S.%L

when .%L is found at init time, it will force a manual fractional
second resolution. Optionally if some data remains after %L it
WILL assume it belongs to a time zone, if so, it will do it
proper resolution too, example of timezone formats:

  +01:00
  -08:30
  Z      (defaults to UTC)

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>