in_tail: handle pending bytes after inotify event (see details)
There are some cases where the capacity of the internal buffer to read
data from a file, is lower than the bytes already available. When Fluent Bit
starts this is not a problem, but once the file is promoted to 'events' mode
through inotify and the event loop system, a read(2) can consume less bytes
than available, then inotify will _not_ notify again, but we need to take care
of that.

The most simple way is to continue read(2)ing until consume all of them, but
this will mean a blocking operation affecting Fluent Bit core, this is _not_
desired.

This patch implements an additional mechanism to detect pending bytes after
an inotify/read operations and use a new signal channel to instruct the core
to consume more bytes while pending data exists. This works in async mode
and do not block the service.

note: this is not an issue before the patch, no data loss exists, just data
is delayed, with this approach we make bytes available ASAP (expected by
users).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>