in_tail: perf improvement when processing file content: from 90.78% to 37.52%
When processing a file content (chunk), the parser routine was invoking memmove()
for every line processed to simplify the buffer management, which ended-up in
some unnnecessary overhead.

This patch makes a better buffer management and now memmove() is only used after
processing the whole number of lines in the buffer.

  before  process_content(): 90.78%
  after   process_content(): 37.52%

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>