in_tail: use dyntag mechanism only
Since in_tail can consume and append an excessive amount of data if
using a basic mode to append records (non dynamic tags), it means the
plugin append data to the input instance buffer until the engine have
to flush which lead to chunks of 5MB or more, sometimes not ideal for
cases when sending the data out to out_forward.

This patch makes in_tail use only dyntag mode. So every chunk size is
not longer than 2.1MB.

note: considering to deprecate that old mode and use dyntag modes
for other plugins.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>