in_mem: add an option in config file to set the collection time interval
Sometimes, it could be useful to set the collection time interval for the mem
measures. For example, if we want to store this log for estimate the average
of mem consumption over a month, it is not usefull to store mem every 1 second
but maybe 60 seconds.
Maybe someone could also be interested in a very accurate mem measurement,
and set collection time less than 1 second with nanoseconds.

By default, collect interval is 1 second and 0 nanosecond.
To set a different interval, set in the .conf file :
Collect_Sec 1
Collect_NSec 0

Signed-off-by: Loris Bererd <bererd.loris@gmail.com>