scheduler: improved scheduling of future retries, decrese fds required
Before this patch, when the scheduler is asked to schedule a future retry,
it used a timer based on a file descriptor. This works pretty fine but on
scenarios where the 'retry' will happen after a long number of seconds, using
a file descriptor is not good, likely the process can reach it limit at some
point.

This patch implements an improved scheduler where now future retries that
happen in the next 10 seconds, are assigned to a timer file descriptor,
otherwise if their expected time to run is longer than that, they are
assigned to a queue and a 'frame checker' lookup the queued retries that
can be promoted every 10 seconds.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>