in_tail: do not mark as rotated files when using relative paths
Sometimes the file names given can be relative paths but when checking
for rotation when "promoting from static to event" mode, the absolute
path is used, generating a 'rotation' mark internally.

For such situations where monitored files are provided with relative
paths, an extra check is done through the check of inode number, this
prevent messy internal rotation because of a user-call error.

note: this is to workaround problems generated by a bad configuration.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>