filter_kubernetes: new 'merge_json_key' option (see details)
This patch introduce a new option called merge_json_key.

When merge_json_log is enabled (set to true), the filter tries to
assume the 'log' field from the incoming message is a JSON string
message and make a structured representation of it at the same
level of the 'log' field in the map.

Now if merge_json_key is set (a string name), all the new structured fields
taken from the original 'log' content are inserted under the new key.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>