env: fix insertion of non variable parts in result
When there is a variable used in the configuration with some text
following it, like `${VAR}-additional-text`, the additional text is not
carried over in the result. This is caused by accidentally overwriting
the variable holding the length of the original string and not copying
the remaining parts of the value into the buffer.

Signed-off-by: Dominik del Bondio <dominik.del.bondio@bitextender.com>
Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>