meta: new support to add meta commands in the configuration (details below)
The configuration files now allows to specify optional 'meta' commands, this
is a way to implement specific actions based in a configuration key.

A meta command format is like this:

  @CMD PARAMETERS

Each meta is prefixed with a '@', then the following is the command name, in
the example is 'CMD' and then after one space the command parameters.

Meta commands are hard-coded commands and at the moment there are two
implemented:

  - SET => @SET key=val: set a local environment variable for the
    configuration runtime.

  - INCLUDE => @INCLUDE file: include a configuration file for the
    current context being created.

note: INCLUDE is an inherit command coming from Monkey runtime configuration
system and is not processed by Fluent Bit.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>