parser: new Time_Offset feature (#326)
Some log entries that don't contain a timezone will not come in UTC format,
so the parser can do a wrong interpretation of the correct UTC time for
each record.

The following patch introduce a new configuration key for the parsers
called Time_Offset, This value specify an optional UTC offset that the
parser can apply to format the proper time in UTC.

Time_Offset takes values in the following format:

 - +0000
 - -0600
 - ...

Example:

 [PARSER]
     Name        test
     Format      json
     Time_Key    time
     Time_Format %Y-%m-%d %H:%M:%S
     Time_Offset -0600
     Time_Keep   On

As of this patch, the JSON backend is the only aware about this new
feature (Regex backend will get this shortly).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>