in_tail: new 'skip_long_lines' option (#370)
By default when there is no more room in the internal buffer to read a
complete line, the file is not longer monitored. There are situations
where users wants to just skip long lines and keep processing other
lines.

This patch introduce a new option to in_tail called 'skip_long_lines',
when the buffer is full and a line has not been processed due to a
missing \n, this option makes to discard that current line content
and process other new lines that fits into the buffer size.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>