in_tail: validate static file status on promotion (#409)
the plugin have two main stages: static and events. On static mode
it parse content files when it start, but then when it reach EOF it
promote the files to 'events', meaning that will use the inotify
mechanism to get file modifications, rotations, etc.

If a monitored file in static mode is deleted, the promotion to events
did not validate properly, upon error the expected behavior is to
unregister the problematic file. This patch fix the issue.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>