http_client: fix buffer new_size calculation
The prior calculation was incorrectly shrinking the buffer when it
reached its limit, leading to data corruption.

Here's an example of what the buffer was resizing to (with added debug
statement):

[2017/11/02 10:56:00] [ info] [filter_kube] testing connectivity with
API server...
resizing http buffer to 28672
resizing http buffer to 4096
resizing http buffer to 28672
resizing http buffer to 4096
Segmentation fault (core dumped)

This is for a max_size of 32768.

Signed-off-by: James Ravn <james@r-vn.org>