input: fix collector start (#416)
Collectors are linked to a global configuration and also to the input
instance where they belongs to. When starting a collector the ID was
compared againts the global list instead of the list in the input instance.

This patch fix the collector lookup by ID using the proper list head.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>