filter_parser: new 'unescape_key' feature (#278)
there are cases where the key record to be parsed comes in a escaped
way, as an example consider Docker and an application that writes
logs in JSON format. Docker is not aware about application logs format
so it just escape the nested record, e.g:

 1. Application output:

    {"key": 123}

 2. Docker Log:

    {"stream": "stdout", "log": "{\"key\": 123}"}

If we wanted to process 'log' key using the JSON parser, it will consider
this as a string and not as a JSON Map.

The new 'unescape_key' option will unescape the key value producing an output
that the parser can deal with:

 3. Fluent Bit + Filter Parser + Unescape_Key ON:

    pre-process: {"stream": "stdout", "log": {"key": 123}}
    pos-process: {"stream": "stdout", "key": 123}

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>