filter_parser: fix reserve_data, add preserve_key
Fix reserve_data which was segfaulting due to multiple parser patch.
This was due to not handling the exit of the parse loop to preserve
existing records.

Add `preserve_key` option so we can also choose to preserve the original
key field, along with `reserve_data`.

Add tests to cover the various scenarios, including time key parsing.