scheduler: new API to trigger callbacks at specific intervals of time
This patch introduce 3 new API functions that can be used to schedule
callbacks functions to be triggered at certain "timeout" time. The
functions registered are called only once and not in a repetitive way.

 - flb_sched_timer_cb_create() : create the timer and associate callback
 - flb_sched_timer_cb_disable(): un-register notifications
 - flb_sched_timer_cb_destroy(): destroy timer

The 'user' only need flb_sched_timer_cb_create(), the other two functions
are used internally by the scheduler.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>