filter_kubernetes: new annotation properties feature
Applications log messages in different ways, not all of them do it
JSON, so likely when the application generate a log message this
can have different formats like json, apache log, mysql log and so on.

When doing log processing the ideal case is to have a structured
log message, but if this does not have a structure it makes things
hard once the logs are stored in a database and we want to do data
analysis.

The following patch aims to improve Fluent Bit Kubernetes Filter
allowing applications to 'suggest' special configuration that the
log processor can use for a better log processing through Annotations:

The current property is supported at the moment:

 - logging.parser: suggest a pre-defined Fluent Bit parser to process
                   the Pod log content

This property must be set in the annotations, specifically in the Pod
spec when the Pod is created (not annotated later). In addition to let
Fluent Bit accept this suggestion the filter must have the configuration
option 'k8s-logging.parser' turned on.

Consider the following Pod definition as an example:

----start----
  apiVersion: v1
  kind: Pod
  metadata:
    name: apache-logs
    labels:
      app: apache-logs
    annotations:
      logging.parser: apache
  spec:
    containers:
    - name: apache
      image: edsiper/apache_logs
-----end-----

the image edsiper/apache_logs simulate Apache logs and generate
one per second, e.g:

 53.245.228.52 - - [09/Jan/2018:20:32:40 +0000] "GET /carminic HTTP/1.0" 500 2216
 13.79.69.247 - - [09/Jan/2018:20:32:41 +0000] "GET /burucha HTTP/1.0" 404 2216
 5.72.89.4 - - [09/Jan/2018:20:32:42 +0000] "GET /Cassell HTTP/1.0" 404 2216
 248.154.245.175 - - [09/Jan/2018:20:32:43 +0000] "GET /biochemica HTTP/1.0" 200 2216
 218.46.130.79 - - [09/Jan/2018:20:32:44 +0000] "GET /achtehalber HTTP/1.0" 404 2216

When the log is generated by the application, Docker will trap it and save it in a
JSON map like this:

{"log":"124.219.247.213 - - [03/Jan/2018:19:30:01 +0000] \"GET /grandchildren HTTP/1.0\" 204 2216\n","stream":"stdout","time":"2018-01-03T19:30:01.793484171Z"}

the value of field 'log' is the real application log, the other content is
metadata associated to the log. When Fluent Bit parse this log content
it will converted into the following structure:

{
 "log": "124.219.247.213 - - [03/Jan/2018:19:30:01 +0000] \"GET /grandchildren HTTP/1.0\" 204 2216",
 "stream": "stdout",
 "time": "2018-01-03T19:30:01.793484171Z",
 "host": "124.219.247.213",
 "user": "-",
 "method": "GET",
 "path": "/grandchildren",
 "code": "204",
 "size": "2216",
 "referer": "",
 "agent": "",
 "kubernetes": {
     "pod_name": "apache-logs",
     "namespace_name": "default",
     "pod_id": "7db2f883-f0bc-11e7-bbd7-0800272b9d6d",
     "labels": {
        "run": "filter-kube-apache"
     },
     "annotations": {
        "logging.parser": "apache"
     },
     "host": "minikube",
     "container_name": "filter-kube-apache",
     "docker_id": "0111c6c6da660f748ee84e819e02393afcca0b301a8aec64525decb4fac2cb77"
 }
}

More properties will be added, for now this looks a good start
to solve the unstructured log problem from a Pod definition level.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>