out_forward: fix memory leak on broken connections
If the option Time_as_Integer is enabled and the destination server
is not available or it closes the TCP connection, the plugin was
leaking the new data buffer composed to deal with old Forward protocol.

This patch makes sure to release the buffer before to return and ask for
the retry.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>